# public/swagger.yaml
openapi: 3.0.0
info:
  title: 'Therapeutic Radiesthesia API'
  description: '`API para demonstrar endpoints básicos do sistema`<br><ul><li>Essa documentação está sendo desenvolvida gradualmente, todos os endpoints estão passando por uma revisão.</li><li>Os endpoints que precisam estar com a autorização estão com um cadeado indicando o uso.</li><li>Para gerar o token, utilize o endpoint login.</li><li>Endpoints marcados como acesso ADMIN.</li></ul>'
  contact:
    name: 'Paulo Henrique'
    url: 'https://api.conect.app'
    email: webmaster@multidesk.io
  version: 1.0.0
servers:
  -
    url: 'https://api.multidesk.io'
    description: 'Servidor online'
  -
    url: 'http://localhost:8000'
    description: 'Servidor local'
paths:
...
  /api/v1/dashboard/appointments:
    get:
      tags:
        - Usuários
      summary: 'Relatórios de compromissos, cancelamentos, anamneses e retornos.'
      description: 'Este endpoint retorna relatórios detalhados de compromissos, cancelamentos, anamneses e retornos do usuário autenticado. O relatório pode ser gerado por ano, mês, semana ou dia, dependendo dos parâmetros passados.'
      operationId: c34783cc58a30ff9f2c9ec4217bd75af
      parameters:
        -
          name: type
          in: query
          description: 'O tipo de relatório. Pode ser "annual", "monthly", "weekly", "daily" ou "compareWithLastWeek". compareWithLastWeek não respeita o campo start e end, pega a semana passada e a semana atual.'
          required: true
          schema:
            type: string
            enum:
              - annual
              - monthly
              - weekly
              - daily
              - compareWithLastWeek
        -
          name: start
          in: query
          description: 'A data de início no formato YYYY-MM-DD. Exigido para relatórios que não sejam "compareWithLastWeek".'
          required: false
          schema:
            type: string
            format: date
        -
          name: end
          in: query
          description: 'A data de fim no formato YYYY-MM-DD. Exigido para relatórios que não sejam "compareWithLastWeek".'
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 'Relatório gerado com sucesso'
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { date: { description: 'Data ou intervalo de tempo no relatório (varia de acordo com o tipo de relatório)', type: string }, appointments: { description: 'Número de compromissos', type: integer }, cancelled: { description: 'Número de compromissos cancelados', type: integer }, anamneses: { description: 'Número de anamneses', type: integer }, return: { description: 'Número de retornos', type: integer } }
                  type: object
        '400':
          description: 'Erro de validação'
        '401':
          description: 'Token inválido ou ausente'
        '500':
          description: 'Erro interno do servidor'
      security:
        -
          bearerAuth: []
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...